First move of Telemetry item except AppRating Review Classes
