+++ App/iOS/Delegates/AppDelegate.swift
@@ -18,6 +18,7 @@ import BraveCore
 import Combine
 import Brave
 import BraveVPN
+import BraveTelemetry
 import RuntimeWarnings
 import BraveNews
 import BraveTalk
