+++ Client/Frontend/Settings/SettingsViewController.swift
@@ -17,6 +17,7 @@ import BraveWallet
 import BraveUI
 import BraveVPN
 import BraveNews
+import BraveTelemetry
 
 extension TabBarVisibility: RepresentableOptionType {
   public var displayString: String {
