+++ Sources/BraveTelemetry/ReferralData.swift
